This commit automatically determines the season of the test.
During the GW call on Nov 2, 2021, we determined season numbers would be the year and month. E.g., a test in November of 2021 would get season 2111. The changes here use the datetime library to determine the current year and month, and the included unit tests verify the season is correctly written to the dagmaker.rc file.